+++ packages/firestore/src/local/index_manager.ts
@@ -164,4 +164,10 @@ export interface IndexManager {
     transaction: PersistenceTransaction,
     target: Target
   ): PersistencePromise<IndexOffset>;
+
+  /** Returns the minimum offset for the given collection group. */
+  getMinOffsetFromCollectionGroup(
+    transaction: PersistenceTransaction,
+    collectionGroup: string
+  ): PersistencePromise<IndexOffset>;
 }