+++ packages/firestore-compat/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @firebase/firestore-compat
 
+## 0.1.19
+
+### Patch Changes
+
+- [`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5) [#6307](https://github.com/firebase/firebase-js-sdk/pull/6307) (fixes [#6300](https://github.com/firebase/firebase-js-sdk/issues/6300)) - fix: add type declarations to exports field
+
+- Updated dependencies [[`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5), [`d4b52b612`](https://github.com/firebase/firebase-js-sdk/commit/d4b52b612cf73610c57a3c08a0415ab7b622a70a)]:
+  - @firebase/component@0.5.15
+  - @firebase/firestore@3.4.10
+  - @firebase/util@1.6.1
+
 ## 0.1.18
 
 ### Patch Changes