+++ packages/messaging/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @firebase/messaging
 
+## 0.9.14
+
+### Patch Changes
+
+- [`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5) [#6307](https://github.com/firebase/firebase-js-sdk/pull/6307) (fixes [#6300](https://github.com/firebase/firebase-js-sdk/issues/6300)) - fix: add type declarations to exports field
+
+* [`c66d59c3d`](https://github.com/firebase/firebase-js-sdk/commit/c66d59c3dcfea71bcdb61715f59911dd8a18d717) [#6327](https://github.com/firebase/firebase-js-sdk/pull/6327) (fixes [#6207](https://github.com/firebase/firebase-js-sdk/issues/6207)) - Fix missing FcmOptions.link
+
+* Updated dependencies [[`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5)]:
+  - @firebase/component@0.5.15
+  - @firebase/installations@0.5.10
+  - @firebase/util@1.6.1
+
 ## 0.9.13
 
 ### Patch Changes