+++ packages/performance/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @firebase/performance
 
+## 0.5.10
+
+### Patch Changes
+
+- [`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5) [#6307](https://github.com/firebase/firebase-js-sdk/pull/6307) (fixes [#6300](https://github.com/firebase/firebase-js-sdk/issues/6300)) - fix: add type declarations to exports field
+
+- Updated dependencies [[`2cd1cc76f`](https://github.com/firebase/firebase-js-sdk/commit/2cd1cc76f2a308135cd60f424fe09084a34b5cb5)]:
+  - @firebase/component@0.5.15
+  - @firebase/installations@0.5.10
+  - @firebase/logger@0.3.3
+  - @firebase/util@1.6.1
+
 ## 0.5.9
 
 ### Patch Changes